fix: fix perf bug in `mox compile` (#159)
* fix compile perf bug

fix a perf bug in mox compile. `compile_` was returning the deployer,
which pickles it to the process's stdout. however, this is not necessary
as the result is not used.

on my machine, this brings warm compile of 36 contracts from 5s to 2.2s.

* fix signature

* fix regression -- inspect module uses compile_

add a wrapper for use by multiprocessing which discards the result.