rework: first refacto of config wrapper for moccasin toml
This a first refacto of the config wapper for moccasin but it is still possible to refine it.
Let's say it is a first try and maybe someone can take on the lead to improve it.

- Moved everything inside a dir config with module architecture
- Init file used to allow retrocompatibility and get what we want to use only
- Managed to move Network logic into another file amf make it abstract to implement the rest in config file

Tested with just unit testing. Not done on integration (need info to setup API key)